refactor: co-locate some files for new project creation
