+++ src/bin/commands/new-project/index.ts
@@ -3,9 +3,9 @@ import { basename, dirname, join } from "path";
 import chalk from "chalk";
 import { cli } from "cli-ux";
 import kebabCase from "lodash.kebabcase";
-import type { CliCommandResponse } from "../../types/cli";
-import { execute } from "../../utils/exec";
-import { gitRootOrCwd } from "../../utils/git";
+import type { CliCommandResponse } from "../../../types/cli";
+import { execute } from "../../../utils/exec";
+import { gitRootOrCwd } from "../../../utils/git";
 import { constructApp } from "./utils/app";
 import { newAppImports, newStackImports, newTestImports } from "./utils/imports";
 import { buildDirectory } from "./utils/init";
