+++ src/bin/commands/new-project/template/.editorconfig
