+++ src/bin/commands/new-project/template/.eslintignore
