+++ src/bin/commands/new-project/template/.eslintrc.js
