+++ src/bin/commands/new-project/template/.gitignore.template
