+++ src/bin/commands/new-project/template/README.md
