+++ src/bin/commands/new-project/template/bin/.ignore
