+++ src/bin/commands/new-project/template/cdk.json
