+++ src/bin/commands/new-project/template/jest.config.js
