+++ src/bin/commands/new-project/template/jest.setup.js
