+++ src/bin/commands/new-project/template/script/setup
