+++ src/bin/commands/new-project/template/tsconfig.eslint.json
