+++ src/bin/commands/new-project/template/tsconfig.json.template
