+++ src/bin/commands/new-project/utils/app.ts
