+++ src/bin/commands/new-project/utils/snapshot.ts
