+++ src/bin/commands/new-project/utils/stack.ts
