+++ src/bin/commands/new-project/utils/utils.ts
